requirements.txt
pandas
numpy
scikit-learn
scipy
matplotlib
seaborn
requests
pyyaml
python-dotenv
pyarrow
